# api_spec.json
{"openapi":"3.1.0","info":{"title":"DeliverU API","version":"0.1.0"},"paths":{"/auth/register":{"post":{"tags":["auth"],"summary":"Register","description":"Register a new user with HKUST email and password.","operationId":"register_auth_register_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegisterRequest"}}},"required":true},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TokenResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/auth/login":{"post":{"tags":["auth"],"summary":"Login","description":"Authenticate a user and return a JWT token.","operationId":"login_auth_login_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TokenResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/users/me":{"get":{"tags":["users"],"summary":"Get Profile","description":"Return the authenticated user's profile.","operationId":"get_profile_users_me_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProfileResponse"}}}}},"security":[{"HTTPBearer":[]}]}},"/users/me/profile":{"put":{"tags":["users"],"summary":"Setup Profile","description":"Create or update the user's profile.","operationId":"setup_profile_users_me_profile_put","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProfileSetupRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProfileResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"HTTPBearer":[]}]}},"/users/me/dark-mode":{"patch":{"tags":["users"],"summary":"Toggle Dark Mode","description":"Toggle dark mode for the authenticated user.","operationId":"toggle_dark_mode_users_me_dark_mode_patch","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/DarkModeToggle"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProfileResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"HTTPBearer":[]}]}},"/users/me/deliverer-toggle":{"patch":{"tags":["users"],"summary":"Toggle Deliverer","description":"Toggle deliverer status for the authenticated user.","operationId":"toggle_deliverer_users_me_deliverer_toggle_patch","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/DelivererToggle"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProfileResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"HTTPBearer":[]}]}},"/orders":{"post":{"tags":["orders"],"summary":"Create New Order","description":"Create a new food delivery order.","operationId":"create_new_order_orders_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/OrderCreateRequest"}}},"required":true},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/OrderResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"HTTPBearer":[]}]}},"/orders/my":{"get":{"tags":["orders"],"summary":"Get My Orders","description":"Get all orders placed by the authenticated user.","operationId":"get_my_orders_orders_my_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/OrderResponse"},"type":"array","title":"Response Get My Orders Orders My Get"}}}}},"security":[{"HTTPBearer":[]}]}},"/orders/queue":{"get":{"tags":["orders"],"summary":"Get Order Queue","description":"Get pending orders available for delivery, filtered by user's preferred halls.","operationId":"get_order_queue_orders_queue_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/OrderResponse"},"type":"array","title":"Response Get Order Queue Orders Queue Get"}}}}},"security":[{"HTTPBearer":[]}]}},"/orders/{order_id}":{"get":{"tags":["orders"],"summary":"Get Order Detail","description":"Get details of a specific order.","operationId":"get_order_detail_orders__order_id__get","security":[{"HTTPBearer":[]}],"parameters":[{"name":"order_id","in":"path","required":true,"schema":{"type":"string","title":"Order Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/OrderResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/orders/{order_id}/accept":{"patch":{"tags":["orders"],"summary":"Accept Delivery","description":"Accept an order for delivery.","operationId":"accept_delivery_orders__order_id__accept_patch","security":[{"HTTPBearer":[]}],"parameters":[{"name":"order_id","in":"path","required":true,"schema":{"type":"string","title":"Order Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/OrderResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/orders/{order_id}/pickup":{"patch":{"tags":["orders"],"summary":"Mark Picked Up","description":"Mark an order as picked up from the canteen.","operationId":"mark_picked_up_orders__order_id__pickup_patch","security":[{"HTTPBearer":[]}],"parameters":[{"name":"order_id","in":"path","required":true,"schema":{"type":"string","title":"Order Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/OrderResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/orders/{order_id}/deliver":{"patch":{"tags":["orders"],"summary":"Mark Delivered","description":"Mark an order as delivered to the orderer.","operationId":"mark_delivered_orders__order_id__deliver_patch","security":[{"HTTPBearer":[]}],"parameters":[{"name":"order_id","in":"path","required":true,"schema":{"type":"string","title":"Order Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/OrderResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/orders/{order_id}/cancel":{"patch":{"tags":["orders"],"summary":"Cancel My Order","description":"Cancel an order (only by orderer, only if pending or accepted).","operationId":"cancel_my_order_orders__order_id__cancel_patch","security":[{"HTTPBearer":[]}],"parameters":[{"name":"order_id","in":"path","required":true,"schema":{"type":"string","title":"Order Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/OrderResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/credits/balance":{"get":{"tags":["credits"],"summary":"Get Credit Balance","description":"Get the authenticated user's credit balance.","operationId":"get_credit_balance_credits_balance_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreditBalanceResponse"}}}}},"security":[{"HTTPBearer":[]}]}},"/credits/history":{"get":{"tags":["credits"],"summary":"Get Credit History","description":"Get the authenticated user's credit transaction history.","operationId":"get_credit_history_credits_history_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreditHistoryResponse"}}}}},"security":[{"HTTPBearer":[]}]}},"/qr/decode":{"post":{"tags":["qr"],"summary":"Decode Qr","description":"Decode a QR code from a base64-encoded image.","operationId":"decode_qr_qr_decode_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/QRDecodeRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/QRDecodeResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"HTTPBearer":[]}]}},"/health":{"get":{"summary":"Health","operationId":"health_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}}},"components":{"schemas":{"CreditBalanceResponse":{"properties":{"credits":{"type":"integer","title":"Credits"}},"type":"object","required":["credits"],"title":"CreditBalanceResponse","description":"Schema for returning credit balance."},"CreditHistoryResponse":{"properties":{"transactions":{"items":{"$ref":"#/components/schemas/CreditTransactionResponse"},"type":"array","title":"Transactions"}},"type":"object","required":["transactions"],"title":"CreditHistoryResponse","description":"Schema for returning credit transaction history."},"CreditTransactionResponse":{"properties":{"id":{"type":"string","title":"Id"},"user_id":{"type":"string","title":"User Id"},"amount":{"type":"integer","title":"Amount"},"reason":{"type":"string","title":"Reason"},"order_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Order Id"},"created_at":{"type":"string","format":"date-time","title":"Created At"}},"type":"object","required":["id","user_id","amount","reason","order_id","created_at"],"title":"CreditTransactionResponse","description":"Schema for returning a single credit transaction."},"DarkModeToggle":{"properties":{"dark_mode":{"type":"boolean","title":"Dark Mode"}},"type":"object","required":["dark_mode"],"title":"DarkModeToggle","description":"Schema for toggling dark mode."},"DelivererToggle":{"properties":{"is_deliverer":{"type":"boolean","title":"Is Deliverer"}},"type":"object","required":["is_deliverer"],"title":"DelivererToggle","description":"Schema for toggling deliverer status."},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"LoginRequest":{"properties":{"email":{"type":"string","format":"email","title":"Email"},"password":{"type":"string","title":"Password"}},"type":"object","required":["email","password"],"title":"LoginRequest","description":"Schema for user login."},"OrderCreateRequest":{"properties":{"canteen":{"type":"string","title":"Canteen"},"items":{"items":{"$ref":"#/components/schemas/OrderItemSchema"},"type":"array","title":"Items"},"total_price":{"type":"number","title":"Total Price"},"delivery_hall":{"type":"string","title":"Delivery Hall"},"note":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Note"},"qr_code_image":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Qr Code Image"},"qr_code_data":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Qr Code Data"}},"type":"object","required":["canteen","items","total_price","delivery_hall"],"title":"OrderCreateRequest","description":"Schema for creating a new order."},"OrderItemSchema":{"properties":{"name":{"type":"string","title":"Name"},"qty":{"type":"integer","title":"Qty"},"price":{"type":"number","title":"Price"}},"type":"object","required":["name","qty","price"],"title":"OrderItemSchema","description":"Schema for a single order item."},"OrderResponse":{"properties":{"id":{"type":"string","title":"Id"},"orderer_id":{"type":"string","title":"Orderer Id"},"deliverer_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Deliverer Id"},"status":{"type":"string","title":"Status"},"canteen":{"type":"string","title":"Canteen"},"items":{"items":{"$ref":"#/components/schemas/OrderItemSchema"},"type":"array","title":"Items"},"total_price":{"type":"number","title":"Total Price"},"delivery_hall":{"type":"string","title":"Delivery Hall"},"delivery_preference":{"type":"string","title":"Delivery Preference"},"qr_code_image":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Qr Code Image"},"qr_code_data":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Qr Code Data"},"note":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Note"},"orderer_nickname":{"type":"string","title":"Orderer Nickname"},"deliverer_nickname":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Deliverer Nickname"},"created_at":{"type":"string","format":"date-time","title":"Created At"},"accepted_at":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Accepted At"},"picked_up_at":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Picked Up At"},"delivered_at":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Delivered At"},"cancelled_at":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Cancelled At"}},"type":"object","required":["id","orderer_id","deliverer_id","status","canteen","items","total_price","delivery_hall","delivery_preference","qr_code_image","qr_code_data","note","orderer_nickname","deliverer_nickname","created_at","accepted_at","picked_up_at","delivered_at","cancelled_at"],"title":"OrderResponse","description":"Schema for returning order data."},"ProfileResponse":{"properties":{"id":{"type":"string","title":"Id"},"email":{"type":"string","title":"Email"},"nickname":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Nickname"},"dorm_hall":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Dorm Hall"},"order_times":{"anyOf":[{"items":{"type":"string"},"type":"array"},{"type":"null"}],"title":"Order Times"},"pref_take_order_location":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Pref Take Order Location"},"pref_delivery_habit":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Pref Delivery Habit"},"is_deliverer":{"type":"boolean","title":"Is Deliverer"},"available_return_times":{"anyOf":[{"items":{"type":"string"},"type":"array"},{"type":"null"}],"title":"Available Return Times"},"preferred_delivery_halls":{"anyOf":[{"items":{"type":"string"},"type":"array"},{"type":"null"}],"title":"Preferred Delivery Halls"},"dark_mode":{"type":"boolean","title":"Dark Mode"},"profile_completed":{"type":"boolean","title":"Profile Completed"},"credits":{"type":"integer","title":"Credits"}},"type":"object","required":["id","email","nickname","dorm_hall","order_times","pref_take_order_location","pref_delivery_habit","is_deliverer","available_return_times","preferred_delivery_halls","dark_mode","profile_completed","credits"],"title":"ProfileResponse","description":"Schema for returning user profile data."},"ProfileSetupRequest":{"properties":{"nickname":{"type":"string","title":"Nickname"},"dorm_hall":{"type":"string","title":"Dorm Hall"},"order_times":{"items":{"type":"string"},"type":"array","title":"Order Times"},"pref_take_order_location":{"type":"string","title":"Pref Take Order Location"},"pref_delivery_habit":{"type":"string","title":"Pref Delivery Habit"},"is_deliverer":{"type":"boolean","title":"Is Deliverer","default":false},"available_return_times":{"anyOf":[{"items":{"type":"string"},"type":"array"},{"type":"null"}],"title":"Available Return Times"},"preferred_delivery_halls":{"anyOf":[{"items":{"type":"string"},"type":"array"},{"type":"null"}],"title":"Preferred Delivery Halls"}},"type":"object","required":["nickname","dorm_hall","order_times","pref_take_order_location","pref_delivery_habit"],"title":"ProfileSetupRequest","description":"Schema for initial profile setup."},"QRDecodeRequest":{"properties":{"image":{"type":"string","title":"Image"}},"type":"object","required":["image"],"title":"QRDecodeRequest","description":"Schema for QR code decode request."},"QRDecodeResponse":{"properties":{"decoded_data":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Decoded Data"}},"type":"object","required":["decoded_data"],"title":"QRDecodeResponse","description":"Schema for QR code decode response."},"RegisterRequest":{"properties":{"email":{"type":"string","format":"email","title":"Email"},"password":{"type":"string","title":"Password"}},"type":"object","required":["email","password"],"title":"RegisterRequest","description":"Schema for user registration."},"TokenResponse":{"properties":{"access_token":{"type":"string","title":"Access Token"},"token_type":{"type":"string","title":"Token Type","default":"bearer"}},"type":"object","required":["access_token"],"title":"TokenResponse","description":"Schema for authentication token response."},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"},"input":{"title":"Input"},"ctx":{"type":"object","title":"Context"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}},"securitySchemes":{"HTTPBearer":{"type":"http","scheme":"bearer"}}}}